[mlir/tfg] Avoid name collision for identity node in the const dedupe hoist pass.
Adding identity node with the same name as a constant makes the MLIR TFG export to the GraphDef fail.
Let the name have `identity/` prefix followed by the const name. There is still a chance that this will lead to collision, but the surface is very limited.
It might worth to reconsider and having some per scope unique name, but that feels an overkill at the moment. I left a TODO to address this.

PiperOrigin-RevId: 431991149