Add ContainerMap forward type inference function
Helper for ops with semantics of changing the content of a container,
e.g. changing the batch size of a dataset.

`T[PRODUCT[U1, ..., Un]] -> T[PRODUCT[V1, ..., Vm]]`

where the i'th input is the input container and its type is T, and m is determined and V is obtained from an "map" mapping U -> V.
Both the "map" mapping is parameterized by this factory.
Note that when the "map" function is the identity function, this becomes equivalent to ReplicateInput.

PiperOrigin-RevId: 432285886